Create death first search solution.rs
This problem plays out on a graph where a “virus” moves from node to node, in search of an exit. There are several exits and you have to cut access to these exits by finding the best link to cut each turn.

Learn to store data in a graph data structure. Searching through and updating the graph.